+++ src/CmsBundle/DependencyInjection/Configuration.php
@@ -87,6 +87,12 @@ class Configuration implements ConfigurationInterface
                 ->arrayNode('blocks')
                     ->addDefaultsIfNotSet()
                     ->children()
+                        ->arrayNode('form')
+                            ->addDefaultsIfNotSet()
+                            ->children()
+                                ->scalarNode('view')->defaultValue('OpiferCmsBundle:Block:Content/form.html.twig')->end()
+                            ->end()
+                        ->end()
                         ->arrayNode('login')
                             ->addDefaultsIfNotSet()
                             ->children()
