+++ src/CmsBundle/Resources/config/doctrine/FormBlock.orm.yml
@@ -1,0 +1,11 @@
+Opifer\CmsBundle\Entity\FormBlock:
+  type: entity
+  manyToOne:
+    form:
+      targetEntity: Opifer\FormBundle\Model\FormInterface
+      fetch: EAGER
+      joinColumn:
+        name: form_id
+        referencedColumnName: id
+        nullable: true
+        onDelete: SET NULL
