+++ src/ContentBundle/DependencyInjection/Configuration.php
@@ -365,12 +365,6 @@ class Configuration implements ConfigurationInterface
                                 ->scalarNode('view')->defaultValue('OpiferContentBundle:Block:Content/downloads.html.twig')->end()
                             ->end()
                         ->end()
-                        ->arrayNode('form')
-                            ->addDefaultsIfNotSet()
-                            ->children()
-                                ->scalarNode('view')->defaultValue('OpiferContentBundle:Block:Content/form.html.twig')->end()
-                            ->end()
-                        ->end()
                         ->arrayNode('breadcrumbs')
                             ->addDefaultsIfNotSet()
                             ->children()
