+++ src/ContentBundle/Resources/config/blocks.yml
@@ -23,7 +23,6 @@ parameters:
     opifer.content.video_block.class: Opifer\ContentBundle\Block\Service\VideoBlockService
     opifer.content.cookiewall_block.class: Opifer\ContentBundle\Block\Service\CookieWallBlockService
     opifer.content.downloads_block.class: Opifer\ContentBundle\Block\Service\DownloadsBlockService
-    opifer.content.form_block.class: Opifer\ContentBundle\Block\Service\FormBlockService
     opifer.content.breadcrumbs_block.class: Opifer\ContentBundle\Block\Service\BreadcrumbsBlockService
     opifer.content.avatar_block.class: Opifer\ContentBundle\Block\Service\AvatarBlockService
     opifer.content.search_results_block.class: Opifer\ContentBundle\Block\Service\SearchResultsBlockService
@@ -196,12 +195,6 @@ services:
         arguments: ['@templating', '%opifer_content.downloads_block_configuration%', '@opifer.media.media_manager', '@service_container']
         tags:
             - { name: opifer.content.block_service, alias: downloads }
-
-    opifer.content.form_block:
-        class: '%opifer.content.form_block.class%'
-        arguments: ['@templating', '@service_container', '@opifer.eav.eav_manager', '%opifer_content.form_block_configuration%']
-        tags:
-            - { name: opifer.content.block_service, alias: form }
 
     opifer.content.breadcrumbs_block:
         class: '%opifer.content.breadcrumbs_block.class%'
