+++ src/FormBundle/Model/FormManager.php
@@ -4,6 +4,8 @@ namespace Opifer\FormBundle\Model;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Opifer\EavBundle\Model\SchemaManager;
+use Opifer\FormBundle\Form\Type\PostType;
+use Symfony\Component\Form\FormFactory;
 
 class FormManager
 {
@@ -21,18 +23,21 @@ class FormManager
 
     /**
      * @param EntityManagerInterface $em
+     * @param FormFactory            $formFactory
      * @param SchemaManager          $schemaManager
      * @param PostManager            $postManager
      * @param string                 $class
+     *
      * @throws \Exception
      */
-    public function __construct(EntityManagerInterface $em, SchemaManager $schemaManager, PostManager $postManager, $class)
+    public function __construct(EntityManagerInterface $em, FormFactory $formFactory, SchemaManager $schemaManager, PostManager $postManager, $class)
     {
         if (!is_subclass_of($class, 'Opifer\FormBundle\Model\FormInterface')) {
             throw new \Exception(sprintf('%s must implement Opifer\FormBundle\Model\FormInterface', $class));
         }
 
         $this->em = $em;
+        $this->formFactory = $formFactory;
         $this->schemaManager = $schemaManager;
         $this->postManager = $postManager;
         $this->class = $class;
@@ -64,6 +69,19 @@ class FormManager
         $form->setSchema($schema);
 
         return $form;
+    }
+
+    /**
+     * Create a Symfony Form instance
+     *
+     * @param FormInterface $form
+     * @param PostInterface $post
+     *
+     * @return \Symfony\Component\Form\FormInterface
+     */
+    public function createForm(FormInterface $form, PostInterface $post)
+    {
+        return $this->formFactory->create(PostType::class, $post, ['form_id' => $form->getId(), 'label' => false]);
     }
 
     /**
